Add section annotations that can be used to signal when a section is complete
And if it contains JS, SML or both.
Empty means not decided yet.
JS and/or SML means that it should contain material on that,
but that it is not done yet.
JS0 and SML0 means that it will NOT cover that language.
JS1 and SML1 means that it is DONE with respect to that language.

Annotations are of course subject to change.